[feat] Implementing ExceptionVerifier

+++ ApiFirst/DomainServices/Interfaces/ICustomerService.cs
@@ -1,15 +1,16 @@
 ﻿using DomainModels.Models;
+using System.Collections.Generic;
 
 namespace DomainServices.Interfaces
 {
     public interface ICustomerService
     {
         Customer Create(Customer model);
-        void Update(int id, Customer model);
-        void Delete(int id);
+        void Update(long id, Customer model);
+        void Delete(long id);
         IEnumerable<Customer> GetAll();
-        Customer GetById(int id);
-        void UpdateEmail(int id, string email);
+        Customer GetById(long id);
+        void UpdateEmail(long id, string email);
         Customer GetByCpf(string cpf);
     }
 }
